backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
pydantic==2.6.0
pydantic-settings==2.1.0
requests==2.31.0
slowapi==0.1.9
redis==5.0.1
# AI / ML
torch==2.2.0
transformers==4.37.2
Pillow==10.2.0
numpy==1.26.3
scipy==1.12.0
# Vector DB
qdrant-client==1.7.3
sentence-transformers==2.3.1

